requirements_cr.txt
absl-py==2.3.1
aenum==3.1.17
ale-py==0.8.1
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
cloudpickle==3.1.2
colorlog==6.10.1
commonroad-drivability-checker==2022.2.1
commonroad-io==2022.3
-e git+https://gitlab.lrz.de/tum-cps/commonroad-rl.git@78cbeeadda8ed19abcc7fefd9d0950d5a79515a7#egg=commonroad_rl
commonroad-route-planner==2022.3
commonroad-vehicle-models==3.0.2
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
gast==0.7.0
glfw==2.10.0
google-pasta==0.2.0
greenlet==3.2.5
grpcio==1.80.0
gym==0.21.0
gym-notices==0.1.0
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.11
imageio==2.34.1
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
iso3166==2.1.1
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==6.1.0
Mako==1.3.11
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mujoco==2.3.3
namex==0.1.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.31.0
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
optuna==4.8.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
Polygon3==3.0.9.1
protobuf==3.20.1
pybind11==3.0.3
pydantic==2.13.0
pydantic_core==2.46.0
pygame==2.1.0
Pygments==2.20.0
PyOpenGL==3.1.10
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==15.0.0
rtree==1.4.1
ruamel.yaml==0.19.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Shapely==1.8.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
stable-baselines==2.10.2
stable_baselines3==2.7.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.4.1
torch==2.8.0
tqdm==4.67.3
triangle==20250106
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
zipp==3.23.1
